Left menu scroll bar removal
Left menu had scroll bar on the div tag even when no extra content was available.  This was caused by to much margin on one of the divs.  Customer CSS added to address this.

Also added exclusion for PyCharm IDE folder and alternate format of build folder exlcusion to .gitignore